requirements.txt
Django==5.1.1
django-cors-middleware==1.5.0
django-extensions==3.2.3
djangorestframework==3.15.2


